fix: move ur controller filtering in reconciler
