+++ cmd/kyverno/main.go
@@ -293,7 +293,6 @@ func main() {
 		kyvernoInformer.Kyverno().V1beta1().UpdateRequests(),
 		eventGenerator,
 		kubeInformer.Core().V1().Namespaces(),
-		log.Log.WithName("BackgroundController"),
 		configuration,
 	)
